weather_app_6/requirements.txt

fastapi
uvicorn
httpx
python-dotenv
sqlalchemy
asyncpg
alembic
psycopg2-binary